requirements.txt
finviz~=1.4.4
beautifulsoup4~=4.9.3
requests~=2.25.1
aiohttp~=3.7.4
lxml~=4.6.3
urllib3~=1.26.5
user_agent~=0.1.9
cssselect~=1.1.0
tqdm~=4.61.1
tenacity~=7.0.0
ratelimit==2.2.1
backoff==2.1.2
python-dotenv==1.0.0